Update chrome-debug to work with -core@v2 updates
